Refactor RabbitMQ configuration to transport-specific
