+++ src/Java/myservicebus-rabbitmq/src/main/java/com/myservicebus/rabbitmq/BusRegistrationContext.java
@@ -1,5 +1,15 @@
 package com.myservicebus.rabbitmq;
 
-public interface BusRegistrationContext {
+import com.myservicebus.di.ServiceProvider;
 
+public class BusRegistrationContext {
+    private final ServiceProvider serviceProvider;
+
+    public BusRegistrationContext(ServiceProvider serviceProvider) {
+        this.serviceProvider = serviceProvider;
+    }
+
+    public ServiceProvider getServiceProvider() {
+        return serviceProvider;
+    }
 }
