--- src/Java/myservicebus-rabbitmq/src/main/java/com/myservicebus/rabbitmq/RabbitMqBusRegistrationConfigurator.java
@@ -1,9 +1,0 @@
-package com.myservicebus.rabbitmq;
-
-import com.myservicebus.BusRegistrationConfigurator;
-import java.util.function.Consumer;
-
-public interface RabbitMqBusRegistrationConfigurator extends BusRegistrationConfigurator {
-    void host(String host, Consumer<RabbitMqHostConfigurator> configure);
-}
-
